% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarker.R
\name{findMarker}
\alias{findMarker}
\title{findMarker}
\usage{
findMarker(mat, cluster, cluster_id)
}
\arguments{
\item{mat}{A (m x n) data matrix of gene expression measurements of individual cells with rows representing genes and columns representing cells.}

\item{cluster}{A cluster output (vector) from clustering}

\item{cluster_id}{Id (numeric) cluster group in `cluster`}
}
\value{
A data frame with gene's ordered by their significance.
}
\description{
Identify marker genes in the given expression matrix.
}
