% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassify.R
\name{scClassify}
\alias{scClassify}
\title{Train and test scClassify model}
\usage{
scClassify(exprsMat_train = NULL, cellTypes_train = NULL,
  exprsMat_test = NULL, cellTypes_test = NULL, tree = "HOPACH",
  algorithm = "WKNN", selectFeatures = "limma",
  similarity = "pearson", cutoff_method = c("dynamic", "static"),
  weighted_ensemble = FALSE, k = 10, topN = 50, hopach_kmax = 5,
  pSig = 0.01, prob_threshold = 0.7, cor_threshold_static = 0.5,
  cor_threshold_high = 0.7, parallel = FALSE, ncores = 1,
  verbose = FALSE)
}
\arguments{
\item{exprsMat_train}{A matrix of expression matrix of training dataset}

\item{cellTypes_train}{A vector of cell types of training dataset}

\item{exprsMat_test}{A list or a matrix indicates the expression matrices of the testing datasets}

\item{cellTypes_test}{A list or a vector indicates cell types of the testing datasets (Optional).}

\item{tree}{A vector indicates the method to build hierarchical tree, set as "HOPACH" by default.
This should be one of "HOPACH" and "HC" (using hclust).}

\item{algorithm}{A vector indicates the KNN method that are used, set as "WKNN" by default. This
should be one or more of "WKNN", "KNN", "DWKNN".}

\item{selectFeatures}{A vector indicates the method to select features, set as "limma" by default.
This should be one or more of "limma", "DV", "DD", "chisq", "BI".}

\item{similarity}{A vector indicates the similarity measure that are used, set as "pearson" by default.
This should be one or more of "pearson",  "spearman", "cosine", "jaccard", "kendall", "binomial", "weighted_rank","manhattan"}

\item{cutoff_method}{A vector indicates the method to cutoff the correlation distribution. Set as "dynamic" by default.}

\item{weighted_ensemble}{A logical input indicates in ensemble learning, whether the results is combined by a weighted score for each base classifier.}

\item{k}{An integer indicates the number of neighbour}

\item{topN}{An integer indicates the top number of features that are selected}

\item{hopach_kmax}{An integer between 1 and 9 specifying the maximum number of
children at each node in the HOPACH tree.}

\item{pSig}{A numeric indicates the cutoff of pvalue for features}

\item{prob_threshold}{A numeric indicates the probability threshold for KNN/WKNN/DWKNN.}

\item{cor_threshold_static}{A numeric indicates the static correlation threshold.}

\item{cor_threshold_high}{A numeric indicates the highest correlation threshold}

\item{parallel}{A logical input indicates whether running in paralllel or not}

\item{ncores}{An integer indicates the number of cores that are used}

\item{verbose}{A logical input indicates whether the intermediate steps will be printed}
}
\value{
A \code{list} of results including
\itemize{
\item{testRes}: A list including the test results for each test dataset and base classifier
\item{trainRes}: A list including the training results including the features that are selected and the hierarchical cell type tree
\item{ensembleRes}: A list including the ensemble results and scores for the test dataset if ensemble learning is performed.
}
}
\description{
Train and test scClassify model
}
\author{
Yingxin Lin
}
