% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scClassify_example}
\alias{scClassify_example}
\title{Example Data for scClassify.}
\format{An \code{list} contains the following objects
\describe{
  \item{exprsMat_xin}{Subset of expression matrix Xin et al. pancreas dataset}
  \item{xin_cellTypes}{Xin et al. cell type information}
  \item{exprsMat_wang}{Subset of expression matrix Wang et al. pancreas dataset}
  \item{wang_cellTypes}{Wang et al. cell type information}
}}
\usage{
scClassify_example
}
\description{
An .rda contains two subsets of human pancreas dataset as examples of scClassify algorithm.
}
\keyword{datasets}
